extract suffix type
