+++ core/state/database.go
@@ -787,7 +787,7 @@ func (tds *TrieDbState) savePreimage(save bool, hash, preimage []byte) error {
 	if !save {
 		return nil
 	}
-	return tds.db.Put(trie.SecureKeyPrefix, hash, preimage)
+	return tds.db.Put(dbutils.PreimagePrefix, hash, preimage)
 }
 
 func (tds *TrieDbState) HashAddress(address common.Address, save bool) (common.Hash, error) {
@@ -811,7 +811,7 @@ func (tds *TrieDbState) HashKey(key *common.Hash, save bool) (common.Hash, error
 }
 
 func (tds *TrieDbState) GetKey(shaKey []byte) []byte {
-	key, _ := tds.db.Get(trie.SecureKeyPrefix, shaKey)
+	key, _ := tds.db.Get(dbutils.PreimagePrefix, shaKey)
 	return key
 }
 
@@ -1057,6 +1057,7 @@ func (dsw *DbStateWriter) UpdateAccountData(ctx context.Context, address common.
 		originalData = make([]byte, originalDataLen)
 		original.EncodeForStorage(originalData)
 	}
+	//fmt.Println("core/state/database.go:1060 Update account",address.String(), dsw.tds.blockNr,  addrHash)
 	return dsw.tds.db.PutS(dbutils.AccountsHistoryBucket, addrHash[:], originalData, dsw.tds.blockNr)
 }
